+++ src/design/plone/contenttypes/tests/test_ct_unita_organizzativa.py
@@ -21,6 +21,239 @@ import transaction
 import unittest
 
 
+class TestUOSchema(unittest.TestCase):
+    layer = DESIGN_PLONE_CONTENTTYPES_API_FUNCTIONAL_TESTING
+
+    def setUp(self):
+        self.app = self.layer["app"]
+        self.portal = self.layer["portal"]
+        self.request = self.layer["request"]
+        self.portal_url = self.portal.absolute_url()
+        setRoles(self.portal, TEST_USER_ID, ["Manager"])
+
+        self.api_session = RelativeSession(self.portal_url)
+        self.api_session.headers.update({"Accept": "application/json"})
+        self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
+
+    def tearDown(self):
+        self.api_session.close()
+
+    def test_behaviors_enabled_for_uo(self):
+        portal_types = api.portal.get_tool(name="portal_types")
+        self.assertEqual(
+            portal_types["UnitaOrganizzativa"].behaviors,
+            (
+                "plone.namefromtitle",
+                "plone.allowdiscussion",
+                "plone.excludefromnavigation",
+                "plone.shortname",
+                "plone.ownership",
+                "plone.publication",
+                "plone.categorization",
+                "plone.basic",
+                "plone.locking",
+                "plone.leadimage",
+                "volto.preview_image",
+                "plone.relateditems",
+                "design.plone.contenttypes.behavior.contatti_uo",
+                "design.plone.contenttypes.behavior.argomenti",
+                "plone.textindexer",
+                "design.plone.contenttypes.behavior.additional_help_infos",
+                "plone.translatable",
+                "kitconcept.seo",
+                "plone.versioning",
+                "collective.taxonomy.generated.tipologia_organizzazione",
+            ),
+        )
+
+    def test_uo_ct_title(self):
+        portal_types = api.portal.get_tool(name="portal_types")
+        self.assertEqual(
+            "Unita Organizzativa", portal_types["UnitaOrganizzativa"].title
+        )
+
+    def test_uo_fieldsets(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(len(resp["fieldsets"]), 12)
+        self.assertEqual(
+            [x.get("id") for x in resp["fieldsets"]],
+            [
+                "default",
+                "cosa_fa",
+                "struttura",
+                "persone",
+                "contatti",
+                "correlati",
+                "categorization",
+                "informazioni",
+                "settings",
+                "ownership",
+                "dates",
+                "seo",
+            ],
+        )
+
+    def test_uo_required_fields(self):
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(
+            sorted(resp["required"]),
+            sorted(
+                [
+                    "title",
+                    "competenze",
+                    "tipologia_organizzazione",
+                    "sede",
+                    "contact_info",
+                    "description",
+                ]
+            ),
+        )
+
+    def test_uo_fields_default_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(
+            resp["fieldsets"][0]["fields"],
+            [
+                "title",
+                "description",
+                "image",
+                "image_caption",
+                "preview_image",
+                "preview_caption",
+                "tassonomia_argomenti",
+            ],
+        )
+
+    def test_uo_fields_cosa_fa_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(resp["fieldsets"][1]["fields"], ["competenze"])
+
+    def test_uo_fields_struttura_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(
+            resp["fieldsets"][2]["fields"],
+            [
+                "legami_con_altre_strutture",
+                "responsabile",
+                "assessore_riferimento",
+                "tipologia_organizzazione",
+            ],
+        )
+
+    def test_uo_fields_persone_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(
+            resp["fieldsets"][3]["fields"],
+            ["persone_struttura"],
+        )
+
+    def test_uo_fields_contatti_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(
+            resp["fieldsets"][4]["fields"],
+            ["contact_info", "sede", "sedi_secondarie", "orario_pubblico"],
+        )
+
+    def test_uo_fields_correlati_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(
+            resp["fieldsets"][5]["fields"],
+            # ["documenti_pubblici", "correlato_in_evidenza"], # BBB dovrebbe essere così
+            # ma viene fuori così nei test perché non viene vista la patch SchemaTweaks
+            ["correlato_in_evidenza"],
+        )
+
+    def test_uo_fields_categorization_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(
+            resp["fieldsets"][6]["fields"],
+            # ["subjects", "language"] BBB dovrebbe essere così
+            # ma nei test esce così perché non viene vista la patch di SchemaTweaks
+            ["subjects", "language", "relatedItems", "documenti_pubblici"],
+        )
+
+    def test_uo_fields_informazioni_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(resp["fieldsets"][7]["fields"], ["ulteriori_informazioni"])
+
+    def test_uo_fields_settings_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(
+            resp["fieldsets"][8]["fields"],
+            [
+                "allow_discussion",
+                "exclude_from_nav",
+                "id",
+                "versioning_enabled",
+                "changeNote",
+            ],
+        )
+
+    def test_uo_fields_ownership_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(
+            resp["fieldsets"][9]["fields"], ["creators", "contributors", "rights"]
+        )
+
+    def test_uo_fields_dates_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(resp["fieldsets"][10]["fields"], ["effective", "expires"])
+
+    def test_uo_fields_seo_fieldset(self):
+        """
+        Get the list from restapi
+        """
+        resp = self.api_session.get("@types/UnitaOrganizzativa").json()
+        self.assertEqual(
+            resp["fieldsets"][11]["fields"],
+            [
+                "seo_title",
+                "seo_description",
+                "seo_noindex",
+                "seo_canonical_url",
+                "opengraph_title",
+                "opengraph_description",
+                "opengraph_image",
+            ],
+        )
+
+
 class TestUO(unittest.TestCase):
     """Test that design.plone.contenttypes is properly installed."""
 
@@ -90,42 +323,6 @@ class TestUO(unittest.TestCase):
 
     def tearDown(self):
         self.api_session.close()
-
-    def test_behaviors_enabled_for_uo(self):
-        portal_types = api.portal.get_tool(name="portal_types")
-        self.assertEqual(
-            portal_types["UnitaOrganizzativa"].behaviors,
-            (
-                "plone.namefromtitle",
-                "plone.allowdiscussion",
-                "plone.excludefromnavigation",
-                "plone.shortname",
-                "plone.ownership",
-                "plone.publication",
-                "plone.categorization",
-                "plone.basic",
-                "plone.locking",
-                "plone.leadimage",
-                "volto.preview_image",
-                "plone.relateditems",
-                # "design.plone.contenttypes.behavior.address_uo",
-                # "design.plone.contenttypes.behavior.geolocation_uo",
-                "design.plone.contenttypes.behavior.contatti_uo",
-                "design.plone.contenttypes.behavior.argomenti",
-                "plone.textindexer",
-                "design.plone.contenttypes.behavior.additional_help_infos",
-                "plone.translatable",
-                "kitconcept.seo",
-                "plone.versioning",
-                "collective.taxonomy.generated.tipologia_organizzazione",
-            ),
-        )
-
-    def test_uo_ct_title(self):
-        portal_types = api.portal.get_tool(name="portal_types")
-        self.assertEqual(
-            "Unita Organizzativa", portal_types["UnitaOrganizzativa"].title
-        )
 
     def test_uo_service_related_service_show_only_services(self):
         response = self.api_session.get(self.uo.absolute_url() + "?fullobjects")